refactor(app): rename error recovery "hooks" dir to "utils"
