+++ app/src/organisms/ErrorRecoveryFlows/utils/__tests__/getFailedCommandPipetteInfo.test.ts
