+++ app/src/organisms/ErrorRecoveryFlows/utils/getFailedCommandPipetteInfo.ts
